+++ src/main/java/com/shadow/jse_middleware_service/constants/NewsType.java
@@ -1,5 +1,11 @@
 package com.shadow.jse_middleware_service.constants;
 
+import java.util.Arrays;
+
 public enum NewsType {
-    DIVDEC
+    DIVDEC;
+
+    public static String getNames() {
+        return Arrays.toString(NewsType.values()).replaceAll("^.|.$", "");
+    }
 }